Removed the example usage from GameController and added it to a comment in Json.cs
